refactor(host-initializer): eliminate duplications and move business logic to main runtime
- Moved interactive mode methods from InitCmd to HostInitializerMain as static methods
- Created reusable promptCancel and handlePromptError functions to eliminate duplication
- Extracted handleInteractiveMode method to reduce complexity in InitCmd
- Moved utility methods (hasGitDirectory, hasWorkspaceInitialized, updateGitignore, setupMcpServer, runInteractiveMode, generateInitMessage) to HostInitializerMain
- Added InteractiveConfig interface for type safety
- Improved separation of concerns: business logic in main runtime, UI concerns in command class
- Fixed return type of init method to include created flag
